+++ CHANGELOG.md
@@ -2,7 +2,8 @@
 
 ## Unreleased
 
-* Disable metrics report for the plugin
+* Improve static code diagnostic `always-remove-listener`.
+* Disable metrics report for the plugin.
 
 ## 4.0.0